fix(flux2): add variable to customise cluster domain (#139)
* add new `clusterDomain` variable which defaults to `cluster.local`
* use `clusterDomain` for source-controller `storage-adv-addr` argument
* use `clusterDomain` for `events-addr` argument

NOTES:

The source-controller manifest included a hard-coded advertising
address with `cluster.local`. In some environments, the cluster domain
is modified, which broke the kustomize controller from being able to
resolve the source controller to acquire artifacts.

BREAKING CHANGE:

The `eventsaddr` variable has been removed as it is no longer necessary.
The new `clusterDomain` variable is now used to create a fully-qualified
addresses throughout the deployment manifiests.

Signed-off-by: Deavon M. McCaffery <dmccaffery@users.noreply.github.com>