fix(charts/flux2,crds): crds.annotations got respected now
In the flux2 helm chart there is already the possibility to
set the value `crds.annotations`. Unfortunately, this value
is not used anywhere because the corresponding templates
are missing in the crds. With this merge request I want to
add templating to the generated crds so that this value can
not only be set without meaning, but also so that the
annotations in the crds can actually be customised.

Signed-off-by: roth-wine <git@roth-wine.de>